Modify raw data display and fix bug in download data formatting
This patch increases the number of samples displayed at once on a graph from
1000 to 2000.
It also fixed a bug in data display which prevents data to be displayed if the
size is equal to the maximum number of data displayed at once.
It also fixed the bug that causes distortion in the recorded data.